iptables: add INPUT vs FORWARD warning in synproxy script
People deploying SYNPROXY seems to reuse this example
and seems to miss that this example is for localhost INPUT chain.

The use INPUT chain where they should have used the FORWARD
chain and then they don't understand why it is not working
on their firewall.

Add a BIG warning notice in the top of the script.

Next they are going to miss, that the TCP options need to match
the host behind the firewall.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>