src: tcp_sink, simple program for accept'ing TCP connections
Simple TCP sink program, that TCP listen() on a port, and accept()
any connection, possibly (not-default) write something into the
connection, and the close() it quickly.

Use-case (1) of this simple program is to help verifying the effect of
SO_REUSEPORT.  By enabling SO_REUSEPORT, starting several instances,
and enable writing the PID and conn count back.  A client can then
connect several times, seeing the that PID change and see if the conn
counter distributes the connection some-what fairly across processes /
listeners.

Use-case (2) is to validate if using SO_REUSEPORT can help mitigate
the listen locking problem, which occurs due to SYN flood DDoS-attacks.
With fake-source SYN floods, the 3WHS (3-way-hand-shake) is never
completed, and the application is never woken up.  Thus, for this
use-case, it is simply enough to start up several instances with
SO_REUSEPORT enabled.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>