src: tcp_sink_client, allow to force source port
Allow to set/force the source port we are connecting from.  This
feature is interesting, because it allow us to better control
hitting the same listen_sock (when testing servers SO_REUSEPORT).

Too fast re-bind and re-connect cause the connect() call to fail
with EADDRNOTAVAIL(99): "connect: Cannot assign requested address"
Implemented connect() retries, but only for this case.

Enable with options:
 --source-port or --sport or -s

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>