scapy: tcp_hacks_3WHS_LAST_ACK.py, introduce 2nd SYN to confuse conntrack
Trying to reproduce a conntrack issue, where conntrack is not correctly
tracking the TCP state LAST_ACK.

Option -f for fail-scenario, will add a 2nd (fake) SYN packet
after the server socket is in LAST_ACK state.  The purpose of
this is to confuse conntrack in changing state.

It seems the timing of sending this 2nd SYN influence whether it
is able to reproduce this issue.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>