pktgen: add comment explaining why grep exit code is not enough
We need to test "result" == "" after the grep test, because I'm also
using the function for /proc/net/pktgen/pgctrl which does not contain
a "Result:" line.  Thus, I can reuse this function for pgctrl, by
letting it pass this test.

Additionally interacting with pgctrl is annoying, because it does
not return errors on false input (kernel patch is submitted to
address this), which made any feedback from pgctrl futile.


The exit code is used when writing to the proc file (see
status=$?), but grepping the "Result:" output is still needed.
This what I need to handle:

On some proc-input pktgen does not return an error, but writes
the error message in the proc file.  Thus, we MUST parse/grep "Result".

On some proc-input pktgen only returns and error, and does NOT
update the "result" output.  Thus, the result can contain
a "Result: OK:" from a previous command.  This was not detected
by old pgset().

E.g this new proc write function catches a very common
configuration sequence of input correct dst IP but wrong MAC addr.

See:
 # echo -e "dst 1.2.3.4.5.6" > eth4@0 ; cat eth4@0 | grep Result
 Result: OK: dst_min=1.2.3.4.5.6
 # echo -e "dst_mac in:va:id " > eth4@0 ; cat eth4@0 | grep Result
 -bash: echo: write error: Invalid argument
 Result: OK: dst_min=1.2.3.4.5.6

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>