pktgen: do arithmetic compare when disabling burst
Replace: [[ ${BURST} != 0 ]]  string pattern compare
With:    [[ ${BURST} -ne 0 ]] arithmetic compare

With bash this still works (comparing it as a string), but I guess it
is more correct as you propose to compare it as a number.

Do realize that this cause $BURST to be evaluated as an arithmetic
expression, thus input like "10-10" will also disable bursting.

In bash it's also possible to write as: (( $BURST != 0 )) to be more
explicit that this is an arithmetic expression, and then we don't use
the -eq, -ne etc.

This "-ne" also works with single "["
(although it does not fall for the arithmetic expansion).

Proposed-by: Cong Wang <cwang@twopensource.com>
Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>