src: compiler_test01: verifying when loops get removed by compiler
The compiler usually removes loops when the expressions inside
the loop evaluates to nothing. But I found a case where this does
not happen.  That is when GCC compiler option -fno-strict-overflow
is used and argument types does not match.

The Linux kernel happen to compile with this option enabled.
This can cause the kernel to keep unnessary loops even-though
debugging CONFIG ifdef options have removed the code inside a
debugging loop.

This was seen with GCC:
 gcc version 4.8.3 20140911 (Red Hat 4.8.3-7) (GCC)

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>