src: udp_flood improve doc for Path MTU discover socket option
For mere mortals (that excludes Eric Dumazet) is not known what
IP_MTU_DISCOVER option corresponds to which constant number.

This change list the IP_MTU_DISCOVER options and corresponding numbers
in the help usage section.  It also (in --verbose mode) prints the
corresponding text representation from the number input of --pmtu.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>