src: udp_sink make sure pktgen check have minimum overhead when disabled
The pktgen contents verification step in check_pkt() is costly
and is guarded by a early if (!p->check) for catching when option
is disabled (which is default off).

The problem is that the overhead of calling/invoking check_pkt()
in a tight measurement loop is also costly.  To help compiler and
branch predictor, create an inlined function call performs the
off check (!p->check).

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>