src: udp_sink handle EAGAIN by trying again
Nonblocking operation (via call flag MSG_DONTWAIT) can be enabled
via --dontwait, and was introduced in commit 7154d8f9c2dc ("udp_sink:
add support for DONTWAIT option and sender address checking").

The commit didn't handle if EAGAIN was returned.  This indicated
that UDP RX queue was empty.  On EAGAIN busy "continue" to
receiving packets but count packets as skipped packets.

This can be useful to detected what overload situation that
occur.  Sometimes we want to test/make-sure that the UDP-sink
userspace application is the bottleneck.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>