trafgen: using dynamic increment dinc() in ICMP ping
The implementation of dynamic increment in trafgen, seems a bit
stange and incomplete IMHO.

The dinc() only works on a single byte, while the ping sequence
number I needed to increment was 2-byte/16-bit.

Next each parallel CPU started at the same "start" number.

The blog documentation here[1] was wrong.
 [1] http://www.draconyx.net/articles/trafgen-expression-language.html

Reading the trafgen code[2][3], the arguments to drnd() is:
 drnd(start, stop, stepping)
But only for a single "byte".

[2] set_dynamic_incdec()
 https://github.com/netsniff-ng/netsniff-ng/blob/v0.6.2/trafgen_parser.y#L362
 https://github.com/netsniff-ng/netsniff-ng/blob/master/trafgen_parser.y#L366
[3] "dinc"
 https://github.com/netsniff-ng/netsniff-ng/blob/v0.6.2/trafgen_parser.y#L689

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>