trafgen: icmp03 experimenting with Destination Unreachable
By sending ICMP Destination Unreachable to the kernel,
it is possible to exercise a "deeper" code path.

This test is related to the attack called BlackNurse:
  http://www.blacknurse.dk/

This test goes one-level deeper, by trying to reach a deeper kernel
code path, by (you manually) creating an established connection on
the target host first, and carefully constructing the inner ICMP
IP-header to match established connection.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>