tc: add example scipt for combining MQ and HTB
Example script for how to solve qdisc locking issue when shaping traffic. Can
be used in cases where global rate limiting it not the goal, but instead the
goal is to rate limit customers or services (to something significantly lower
than NIC link speed).

This work is not complete, as it also requires a component that place
the packet on the right CPU to avoid the locking issue, and correct
shaping.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>